relocate save_img() from core.utils.misc to core.data.utils.misc
